fix: managed to mitigate all comp time errors
