+++ translatable_proc/src/data/translations.rs
@@ -2,44 +2,13 @@ use std::collections::HashMap;
 use std::fs::{read_dir, read_to_string};
 use std::sync::OnceLock;
 
-use proc_macro2::{Span, TokenStream};
-use quote::quote;
-use strum::ParseError;
-use syn::LitStr;
 use thiserror::Error;
 use toml::{Table, Value};
 
+use translatable_shared::{Language, NestingType};
+
 use super::config::{SeekMode, TranslationOverlap, load_config};
 use crate::translations::errors::TranslationError;
-
-/// Errors occurring during TOML-to-translation structure transformation
-#[derive(Error, Debug)]
-pub enum TransformError {
-    /// Mixed content found in nesting node (strings and objects cannot coexist)
-    #[error("A nesting can contain either strings or other nestings, but not both.")]
-    InvalidNesting,
-
-    /// Template syntax error with unbalanced braces
-    #[error("Templates in translations should match '{{' and '}}'")]
-    UnclosedTemplate,
-
-    /// Invalid value type encountered in translation structure
-    #[error("Only strings and objects are allowed for nested objects.")]
-    InvalidValue,
-
-    /// Failed to parse language code from translation key
-    #[error("Couldn't parse ISO 639-1 string for translation key")]
-    LanguageParsing(#[from] ParseError),
-}
-
-/// Represents hierarchical translation structure
-#[derive(Clone)]
-pub enum NestingType {
-    /// Nested namespace containing other translation objects
-    Object(HashMap<String, NestingType>),
-    /// Leaf node containing actual translations per language
-    Translation(HashMap<Iso639a, String>),
-}
 
 /// Translation association with its source file
 pub struct AssociatedTranslation {
@@ -78,21 +47,6 @@ fn walk_dir(path: &str) -> Result<Vec<String>, TranslationError> {
     }
 
     Ok(result)
-}
-
-/// Validates template brace balancing in translation strings
-fn templates_valid(translation: &str) -> bool {
-    let mut nestings = 0;
-
-    for character in translation.chars() {
-        match character {
-            '{' => nestings += 1,
-            '}' => nestings -= 1,
-            _ => {},
-        }
-    }
-
-    nestings == 0
 }
 
 /// Loads and caches translations from configured directory
@@ -141,99 +95,6 @@ pub fn load_translations() -> Result<&'static Vec<AssociatedTranslation>, Transl
     Ok(TRANSLATIONS.get_or_init(|| translations))
 }
 
-impl NestingType {
-    /// Resolves a translation path through the nesting hierarchy
-    ///
-    /// # Arguments
-    /// * `path` - Slice of path segments to resolve
-    ///
-    /// # Returns
-    /// Reference to translations if path exists and points to leaf node
-    pub fn get_path(&self, path: Vec<&str>) -> Option<&HashMap<Iso639a, String>> {
-        match self {
-            Self::Object(nested) => {
-                let (first, rest) = path.split_first()?;
-                nested.get(*first)?.get_path(rest.to_vec())
-            },
-            Self::Translation(translation) => path.is_empty().then_some(translation),
-        }
-    }
-}
-
-impl From<NestingType> for TokenStream {
-    /// Converts NestingType to procedural macro output tokens
-    fn from(val: NestingType) -> Self {
-        match val {
-            NestingType::Object(nesting) => {
-                let entries = nesting.into_iter().map(|(key, value)| -> TokenStream {
-                    let key = LitStr::new(&key, Span::call_site());
-                    let value: TokenStream = value.into();
-                    quote! { (#key.to_string(), #value) }
-                });
-
-                quote! {
-                    translatable::internal::NestingType::Object(vec![#(#entries),*].into_iter().collect())
-                }
-            },
-
-            NestingType::Translation(translation) => {
-                let entries = translation.into_iter().map(|(lang, value)| {
-                    let lang = LitStr::new(&format!("{lang:?}").to_lowercase(), Span::call_site());
-                    let value = LitStr::new(&value, Span::call_site());
-
-                    quote! { (#lang.to_string(), #value.to_string()) }
-                });
-
-                quote! {
-                    translatable::internal::NestingType::Translation(vec![#(#entries),*].into_iter().collect())
-                }
-            },
-        }
-    }
-}
-
-impl TryFrom<Table> for NestingType {
-    type Error = TransformError;
-
-    /// Converts TOML table to validated translation structure
-    fn try_from(value: Table) -> Result<Self, Self::Error> {
-        let mut result = None;
-
-        for (key, value) in value {
-            match value {
-                Value::String(translation_value) => {
-                    // Initialize result if first entry
-                    let result = result.get_or_insert_with(|| Self::Translation(HashMap::new()));
-
-                    match result {
-                        Self::Translation(translation) => {
-                            if !templates_valid(&translation_value) {
-                                return Err(TransformError::UnclosedTemplate);
-                            }
-                            translation.insert(key.parse()?, translation_value);
-                        },
-                        Self::Object(_) => return Err(TransformError::InvalidNesting),
-                    }
-                },
-
-                Value::Table(nesting_value) => {
-                    let result = result.get_or_insert_with(|| Self::Object(HashMap::new()));
-
-                    match result {
-                        Self::Object(nesting) => {
-                            nesting.insert(key, Self::try_from(nesting_value)?);
-                        },
-                        Self::Translation(_) => return Err(TransformError::InvalidNesting),
-                    }
-                },
-
-                _ => return Err(TransformError::InvalidValue),
-            }
-        }
-
-        result.ok_or(TransformError::InvalidValue)
-    }
-}
 
 impl AssociatedTranslation {
     /// Gets the original file path of the translation
