+++ translatable_shared/Cargo.toml
@@ -9,4 +9,9 @@ edition = "2024"
 authors = ["Esteve Autet <esteve@memw.es>", "Chiko <chiko@envs.net>"]
 
 [dependencies]
+proc-macro2 = "1.0.94"
+quote = "1.0.40"
 strum = { version = "0.27.1", features = ["derive", "strum_macros"] }
+syn = { version = "2.0.100", features = ["full"] }
+thiserror = "2.0.12"
+toml = "0.8.20"
