+++ translatable_shared/src/lib.rs
@@ -4,5 +4,5 @@ mod nesting_type;
 
 /// Export all the structures in the common
 /// top-level namespace
-pub use crate::languages::Language;
-pub use crate::nesting_type::NestingType;
+pub use crate::languages::{Language, Similarities};
+pub use crate::nesting_type::{NestingType, TransformError};
